ci: Improve coverage collection
The old coverage collection would fail because the tarpaulin version was
outdated. To avoid this in the future, we change the job script to
always build tarpaulin from source. This induces longer CI times, but
given how infrequently CI runs, we can live with that.

Some other improvements made while we're at it:

 * The coverage collection is split into a separate job, to not
   compromise the speed of the rust-latest job.
 * A cobertura report is provided for Gitlab to show in the MR diff.