Use const generics internally
This removes some macro boilerplate by replacing it with const generics.
Specifically, the implementers of the `CFft`/`RFft`/`IFft` traits aren't
separate types anymore but a single type that's generic over `N`.

The thing that prevents us from getting rid of macros entirely, and to
also make the user-facing functions generic, is lack of support for
calculating `N / 2` with a const generic `N` in stable Rust.