Extend and rework unit tests
This patch extends the test coverage with various types of singnals.
The motivation for this was that the original IRFFT implementation was
working for basic singals with a single sine cycle per each block, but
not with anything more complex.

This patch also adds a check comparing the output of RFFT with the
reference implementation from the realfft crate.

Finally, this patch cleans up tests, removing those cases that were
already covered by other tests.

Signed-off-by: Petr Horacek <petr@zlosynth.com>